+++ CHANGELOG.md
@@ -34,6 +34,7 @@
 - [BREAKING] Rename `ProvingServiceError` to `RemoteProverError` (#1004).
 - [BREAKING] Renamed `Note` to `CommittedNote`, and `NetworkNote` to `Note` in the proto messages (#1022).
 - [BREAKING] Limits of store queries per query parameter enforced (#1028).
+- Migrate from `rusqlite` to `diesel` as a database abstraction (#921)
 
 ### Fixes